STEP 1: Create form in the first screen
The user can select thanks to a spinner a country value
between 0 to 7. And by clicking a simple button access
to a DatePickerDialog in order to select date but also
hour. However another widget like TimePickerDialog could be useful be not user friendly. A textView display the data from the widget after the user selection.