Install and configure database with Room and ViewModel
This is only a set up, for now all the features necessary to handle data within a database are set. All libraries like ViewModel, LiveData and Room are installed. The idea is to grant the user to store his favorites product inside it and perform classic CRUD operations. Next step is the implementation of the last features.