Add a new list feature and optimize UI
A new adapter to display a list of favorites items is created. Now the user can see whenever he wants a list of the items he has selected. A great update of the UI is done too. This allow a better user experience since two lists are displayed into the same recyclerView.