STEP 5: Possible to add or delete products from favorites
The user can, in the detail screen, add or delete product object to/from Room database. The FAB icon evolve every time the user click on it. The icon is automatically changed when the displayed product has the same reference with the one stored in Room. Again it helps the user to understand what kind of action he can make. And he can display the list of all the favorite product stored.