Fix bug on the click of the database product list
The app allow to display both products from web service and product from local database in the same list. When the user clicks on an item, it shows the product detail, but it wasn't working from the database list. Because the UI wasn't correctly adjusted to the correct set of data. This is rectified.